requirements.txt
pydantic==1.8.2
pytest==7.2.1
python-dotenv==1.0.0
SQLAlchemy==2.0.23
